+++ mycodo/inputs/sense_hat.py
@@ -3,7 +3,6 @@ import copy
 
 from mycodo.inputs.base_input import AbstractInput
 
-# Measurements
 measurements_dict = {
     0: {
         'measurement': 'temperature',
@@ -77,8 +76,6 @@ measurements_dict = {
     }
 }
 
-
-# Input information
 INPUT_INFORMATION = {
     'input_name_unique': 'PI_SENSE_HAT',
     'input_manufacturer': 'Raspberry Pi Foundation',
@@ -92,7 +89,20 @@ INPUT_INFORMATION = {
                'which include the LPS25H, LSM9DS1, and HTS221.',
 
     'dependencies_module': [
-        ('pip-pypi', 'sense_hat', 'sense-hat==2.2.0'),
+        ('apt', 'git', 'git'),
+        ('bash-commands',
+         [
+             '/var/mycodo-root/env/RTIMULib_installed',
+         ],
+         [
+             'cd /tmp',
+             'git clone https://github.com/RPi-Distro/RTIMULib',
+             'cd ./RTIMULib/Linux/python/',
+             '~/Mycodo/env/bin/python setup.py build',
+             '~/Mycodo/env/bin/python setup.py install',
+             'touch /var/mycodo-root/env/RTIMULib_installed'
+         ]),
+        ('pip-pypi', 'sense_hat', 'sense-hat==2.2.0')
     ],
 
     'interfaces': ['I2C'],